face_authenticator/requirements.txt
face_recognition>=1.3.0
numpy>=1.24.0
flask>=3.0.0
Pillow>=10.0.0
